@@ -5,7 +5,7 @@ publish_to: "none"
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.12.0 <3.0.0"
+  sdk: ">=2.18.0 <4.0.0"
 
 dependencies:
   flutter:
@@ -16,7 +16,7 @@ dependencies:
   get_storage: ^2.1.1
 
   # Network
-  dio: ^5.3.3
+  dio: ^5.7.0
 
   # UI
   shimmer: ^3.0.0
